% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Severity.R
\name{SeverityForAllCountries}
\alias{SeverityForAllCountries}
\title{Severity for all countries}
\usage{
SeverityForAllCountries(infile)
}
\description{
This is reading the excel file made of many sheets and returns a single long dataset for all countries
}
\examples{
SeverityForAllCountries()
}
