% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Severity.R
\name{SeverityPerCountry}
\alias{SeverityPerCountry}
\title{Severity per country}
\usage{
SeverityPerCountry(infile, country)
}
\description{
This is reading the excel file made of many sheets and returns a single long dataset for the selected country
}
\examples{
SeverityPerCountry()
}
